Cleanup jslint and refine the header check
This should probably have been two commits, but it cleans up the
jshint errors that were hanging around and cleans up the
getCompression file to make it clearer. Removed the q calculation
because it never seems to be used and wasn't working. It also introduces
some tool/util functions for use internally.

Namely: isDefined which checks if an item is defined and
.not() which returns the inverse of any boolean passed into it.