Adds more tests and refactors a tiny bit
This adds the tests needed to hit the 80% coverage threshold.
There are still some weak tests and the tests around serving
compressed static files are missing still... but this will do for
now. Bed time.

I'd also like to see if there is a way to break down the router
more. And specifically to look at pushing some of the serving
into async threads.