Cleans up module syntax for consistency
These were the only modules not declaring their export as a
variable at the top of the page. I like that syntax... so sticking
to it.

The only exception to the rule is the tools module. All of the
functions in it so far are crazy simple and fit in the export
statement nicely. So that is how it will stay for now.