Updates dependencies, simplifies parsing, and adds tests
First off this commit removes the explicit dependency on
event-state. This module was used back in the ancient days
when we were not using harken. But since harken includes it
already and provides a .required of its own we don't need
event-state any more.

Second, this commit breaks out the form string parsing into its
own module. This allows it to be tested directly, and updated
directly. Simplifies the parser.js file a bit and makes things
a little more tidy.

Lastly, adds tests for the parseForm module. This was what spurred
the change above. Because it had the last line of code in parser.js
that was not tested. Pulling it out and testing it means that both
of them are now 100% covered by unit tests.

One weird side note... I am seeing some inconsistent test failures
in the static routing stuff. Not sure why and will investigate
further in the future.