Adds support for the x-powered-by header
By default this header is not sent and that continues to be the
case. But if you want to send it you now may. It lives in the
security section of the config object. DO NOT use this unless you
are trying to be funny and/or understand the security ramifications.

The default is secure. Changing from it is less so.