Updates the readme and implements X-XSS-Protection
There is now an option to setup the X-XSS-Protection header.
It defaults to the safest versions for the requesting browser
avoiding a vulnerability in IE < 9 and enabling for everything
else. Not many browsers support this header, but it is has no
effect on the ones that don't support it.

Updated the readme with a breakdown of the config object as it
stands for the next release.