Adds noCache function to the response
Completely revises the internals of the noCache module so that
if you want to be more precise then the nuclear option of shutting
down all caching you can do so in a more surgical way. With the
res.noCache function. Takes true false to determine if etags are
also blown away.