Fixes a bug introduced by the switch to the new UserAgent system
One of the IE11 strings in use was the backwards compat IE 8 string
and the chrome string was not being used because it wasn't
actually implemented. This is all fixed. There is also a function
that returns the user-agent header and tries to take into account
the various capitalizations so that it has a value no matter how
the browser sends it across the wire.