Initial pass through the CSP code
It is now working at a base level of functionality with no real
documentation (no documentation at all!). But it is passing the
test suite from Helmet and that means it is functional.

Still need to get smart default policies in place and then look
at performance improvements and code simplification.