Improves performance of CSP header and improves quality of tests
The tests were intermittently failing because of a timing issue.
They are now process.nextTicked to prevent that in the future and
passing 100% of the time.

CSP was generating the header on evry request but there are really
only 3 possible values for a given config. Since config can only be
changed at server startup it makes sense to cache these three values
and return them faster. This implements caching.