Adds tons of documentation and cleans up the implementation
The initial implementation had a bunch of weaknesses. This is much
better, allowing greater flexibility, more control, and better
settings. The default is still `false` for te webSocket server but
that could change in the future if enough people are using it to
justify the switch.

The documentation walks through the use of all the settings and the
two different event systems supported.