Adds basic support for http2 spdy
First rough run at http2 support. Right now you have to pass it
an http2 server factory like the `http2` module on npm. Which is
what I am using for testing at this point.

This should probably become a flag in config. It also requires
that you pass it a key and cert for it to work.

Not sure exactly how this plays with heroku or other hosted systems