Finishes the implementation of adhoc server types
Now you can pass in any server type you want and a config object
for it and be up and running. Adds tests around this for http2
module and the spdy module as they are the primary ones we never
want to break.

It adds some test certificates for testing the above since they
both need them.

It also cleans up some tests that were pending and not needed and
a couple of code paths that were never executed.