Fixes a bug in the CSP code
It isn't completely documented but the CSP code expects an array of
directives not a string, and if you pass a string and a version of
firefox < 23 and > 4 visits your site everything blows up in sorrow.

This code will allow you to use either a string or an array without
crashing your site.