This fixes some isues with error handling and one that broke all
The message format was incorrect in the send module of monument. WS
requires that the payload be a string or buffer and we were passing
it an object. I am really not sure how that slipped through since it
would never have worked at all... but it did.

This corrects the format of the message and changes the error
handling so that it works as expected. Only firing the error event
and logging when it fails instead of everytime with undefined as the
payload.