Implements brotli compression
This commit implements brotli compression using the brotli npm
library. This may not be ideal because it is a js port of brotli
instead of the actual c code. However the c port seems to have less
then ideal memory usage at the moment. If there is a major
degredation in performance we may want to look at creating a better
c extension for it for node.

There is also some cleanup of the setup code that clears previously
compressed files. It has been moved to an array of file globs and
an array iterator instead of hardcoding each one. This is no more
effecient (potentially less so) but it happens at startup only and
provides cleaner code with easier expansion going forward.