Adds the .npmignore file
Excludes the following directories:
docs
coverage
test_stubs
bin

and the following dot files:
.gitignore
.codeclimate.yml
.eslintrc
.editorconfig
.linthub.yml

These files are not needed for the module's use and are really
development concerns. So there was no need for them to be loaded
with every install of monument.

This should cut the installed size by about 1.9MB maybe a little
more.