Adds the optional C based deps for ws library
The optional deps utf-8-validate and bufferutil improve
performance for the web sockets part of monument substiantially.

Well at least that is what the documentation for ws says here:
https://www.npmjs.com/package/ws#opt-in-for-performance

They are added as optional dependencies so that the install will
continue even if the install of them fails. Important since they
are binary add ons and they have to be compiled for each platform
as part of the install.