Adds the routes API
Closes #166

This is the big first commit for the route api. It switches most
of the places over to it. There may be a few places that need to
still be switched over and a search through the code base will be
needed to ensure that they are all found.

The other thing that needs to be done is the tests are still in
need of completion. Mainly this has to do with the API for removing
a route, since their is no where that does that in the code base
itself.

Other then that the server is working with the new API in place and
can run. So this is probably pretty close to what it is going to
end up looking like.