Added a note and started the work of getting precompile
So there are a bunch of things that aren't clear about precompile
for tepmlate engines other then dotjs. For instance... should the
templates be loaded into memory? Written out to files? Placed in a
store of some sort?

This commit will set all that up and allow users to provide a custom
precompile function that handles all of the precompile for them.