updates all dependencies and works on statsd
Using this docker image:
https://hub.docker.com/r/kamon/grafana_graphite/
to setup a local statsd and make sure data is being
logged.

Had to bypass the increment method on the statsd
client because it never seemed to be sent. So just
used the calls that it was a convenience for instead.

We are logging out the response status of every request
along with the amount of time that it takes to respond
to each request(with the path as a place to filter).
This should be a pretty good starting place. May expand
in future releases to other metrics.

Updated the following deps:
  - raw-body        ~2.1.4  →  ~2.2.0
  - mocha           ~3.1.2  →  ~3.2.0
  - bufferutil      ~1.2.1  →  ~1.3.0
  - utf-8-validate  ~1.2.1  →  ~2.0.0
  -  busboy      ~0.2.12  →   ~0.2.14
  - platform     ~1.3.0  →    ~1.3.3
  - coveralls  ~2.11.11  →  ~2.11.15
  - http2        ~3.3.2  →    ~3.3.6
  - istanbul     ~0.4.4  →    ~0.4.5
  - spdy         ~3.4.0  →    ~3.4.4