Fixes the format of the etag in the test
The etag format has changed between version of the
etag library we are using. This fixes the test so that
it will pass correctly.

There are some end user ramifications to this. The biggest
being that anything that is cached by etag will be cache
broken with this update and re-downloaded. Other then that
should be smooth sailing.