working on getting tests passing
lots of changes... not totally happy with some of them
but I need a better way and haven't found it yet.

The main culprit in that regard is the config object being
modified in the router now. I don't like it but haven't thought
of a better way to handle ensuring a default templating engine
is in place by the time it is needed.