Breaks out router tests into smaller files
This will facilitate better isolation, and easier tracking
down of what is making things flaky here.

It also makes changes to the route setup process
that prevent fast running tests from hitting an
unprocessed static routes list. This was async before
and caused issues when the static routes tests are
run in isolation.

It also adds the illustrious flaky.sh which can be passed
a file to test or nothing in which case it runs the
mocha script from package.json. It excutes either of these
50 times back to back to give a good test for flakiness in
tests.

Another big shift is changing the names of all the test files
in preperation for the move to ava and to minimize config.