Removes templating completely...
Tepmlates and choice of template engine is something that is
better left to individual developers then us. This simplifies our
code, and allows for maximum flexibility. If you want to use
React, do it. If you want to use dot, do it.

A corresponding change will be made to the CLI once this is all
working that will allow you to choose from common template
systems and drop the correct one for you.